Safari-611.1.10/LayoutTests/js/dom/Promise-resolve-expected.txt

Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS thisInFulfillCallback === secondPromise is false
PASS thisInFulfillCallback === window is true
PASS result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
